+++ packages/rum-core/src/index.ts
@@ -1,0 +1,12 @@
+export { RumUserConfiguration, RumPublicApi, makeRumPublicApi } from './boot/rumPublicApi'
+export { ProvidedSource } from './domain/rumEventsCollection/error/errorCollection'
+export {
+  RumEvent,
+  RumActionEvent,
+  CommonProperties,
+  RumErrorEvent,
+  RumViewEvent,
+  RumResourceEvent,
+  RumLongTaskEvent,
+} from './rumEvent.types'
+export { startRum } from './boot/rum'