+++ scripts/bs-wrapper.js
@@ -5,7 +5,7 @@ const request = require('request')
 
 const AVAILABILITY_CHECK_DELAY = 30_000
 const RUNNING_BUILDS_API = `https://${process.env.BS_USERNAME}:${process.env.BS_ACCESS_KEY}@api.browserstack.com/automate/builds.json?status=running`
-const COMMAND = `yarn ${process.argv.slice(2).join(' ')}`
+const COMMAND = process.argv.slice(2).join(' ')
 const RETRY_DELAY = 30_000
 const MAX_RETRY_COUNT = 3
 
@@ -54,12 +54,14 @@ function hasRunningBuild() {
 function runTests() {
   return new Promise((resolve) => {
     let logs = ''
+    let isKilled = false
     const current = exec(COMMAND)
     current.stdout.pipe(process.stdout)
     current.stdout.on('data', (data) => {
       logs += data
 
-      if (hasSessionCreationFailure(logs)) {
+      if (!isKilled && hasSessionCreationFailure(logs)) {
+        isKilled = true
         current.kill('SIGTERM')
       }
     })
@@ -67,7 +69,7 @@ function runTests() {
     current.on('exit', (code) => {
       if (code === 0) {
         resolve(TEST_STATUS_SUCCESS)
-      } else if (hasSessionCreationFailure(logs)) {
+      } else if (isKilled || hasSessionCreationFailure(logs)) {
         resolve(TEST_STATUS_RECOVERABLE_FAILURE)
       } else {
         resolve(TEST_STATUS_DEFINITIVE_FAILURE)