+++ webpack.base.js
@@ -1,4 +1,7 @@
+const path = require('path')
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin')
+
+const tsconfigPath = path.join(__dirname, 'tsconfig.base.json')
 
 module.exports = (mode) => ({
   mode,
@@ -9,11 +12,18 @@ module.exports = (mode) => ({
         test: /\.ts$/,
         loader: 'ts-loader',
         exclude: /node_modules/,
+        options: {
+          configFile: tsconfigPath,
+          onlyCompileBundledFiles: true,
+          compilerOptions: {
+            module: 'es6',
+          },
+        },
       },
     ],
   },
   resolve: {
     extensions: ['.ts', '.js'],
-    plugins: [new TsconfigPathsPlugin()],
+    plugins: [new TsconfigPathsPlugin({ configFile: tsconfigPath })],
   },
 })